sokoban.sed v0.2: remove lame #r comment prefix
When I was writing this sed script in Vim, the syntax highlight got lost.
Prefixing all comments with `#r` instead of `#` solved the highlighter problem.

Hours later I've discovered that the culprit was a strange character in the
code. I have removed the char, but forgot to remove the silly `#r` prefix.

Now it's done.